REI refactors again + code cleanup

+++ astromine-core/src/main/java/com/github/mixinors/astromine/client/rei/electricsmelting/ElectricSmeltingCategory.java
@@ -24,9 +24,19 @@
 
 package com.github.mixinors.astromine.client.rei.electricsmelting;
 
+import java.text.DecimalFormat;
+import java.util.List;
+
+import com.google.common.collect.Lists;
+
 import com.github.mixinors.astromine.client.rei.AMRoughlyEnoughItemsPlugin;
 import com.github.mixinors.astromine.registry.common.AMBlocks;
-import com.google.common.collect.Lists;
+
+import net.minecraft.text.TranslatableText;
+
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+
 import me.shedaniel.math.Point;
 import me.shedaniel.math.Rectangle;
 import me.shedaniel.rei.api.client.gui.widgets.Widget;
@@ -34,12 +44,6 @@ import me.shedaniel.rei.api.client.gui.widgets.Widgets;
 import me.shedaniel.rei.api.common.util.EntryStacks;
 import me.shedaniel.rei.plugin.client.categories.cooking.DefaultCookingCategory;
 import me.shedaniel.rei.plugin.common.displays.cooking.DefaultCookingDisplay;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-import net.minecraft.text.TranslatableText;
-
-import java.text.DecimalFormat;
-import java.util.List;
 
 @Environment(EnvType.CLIENT)
 public class ElectricSmeltingCategory extends DefaultCookingCategory {
@@ -56,7 +60,7 @@ public class ElectricSmeltingCategory extends DefaultCookingCategory {
 		widgets.add(Widgets.createRecipeBase(bounds));
 		widgets.add(Widgets.createResultSlotBackground(new Point(startPoint.x + 61, startPoint.y + 19)));
 		if (display instanceof ElectricSmeltingDisplay electricSmeltingDisplay)
-			widgets.addAll(AMRoughlyEnoughItemsPlugin.createEnergyDisplay(new Rectangle(bounds.getX() + 10, bounds.getCenterY() - 23, 12, 48), electricSmeltingDisplay.getEnergyRequired(), false, (long) (cookingTime / 10 * 500)));
+			widgets.addAll(AMRoughlyEnoughItemsPlugin.createEnergyDisplay(new Rectangle(bounds.getX() + 10, bounds.getCenterY() - 23, 12, 48), electricSmeltingDisplay.getEnergyRequired(), false, (int) (cookingTime / 10 * 500)));
 		widgets.add(Widgets.createLabel(new Point(bounds.x + bounds.width - 5, bounds.y + 5), new TranslatableText("category.astromine.cooking.time", df.format(cookingTime / 40.0F))).noShadow().rightAligned().color(0xFF404040, 0xFFBBBBBB));
 		widgets.add(Widgets.createArrow(new Point(startPoint.x + 27, startPoint.y + 18)).animationDurationTicks(cookingTime / 3));
 		widgets.add(Widgets.createSlot(new Point(startPoint.x + 4, startPoint.y + 19)).entries(display.getInputEntries().get(0)).markInput());
